chore(TKC-1458): move TCL files to pkg/tcl
